oops, dont enable appstream by default for mock
packages in the appstream repo generally do not rely on other packages in that repo as dependencies.
ie they are generally able to be built standalone without the appstream repo.
very rarely do we need to enable the appstream repo for building. Furhtermore
the davinci resolve runtime can cause some build issues. We should keep appstream explicitly
disabled in the mock files unless in the rare case something from it is needed.